Moved ServiceDiscovererUtils to dns-discovery-netty module
